third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 26.36799999891082 but got 27.478666666866047
Harness: the test ran to completion.
